+++ src/Managers/Assistants/SortAssistant.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\View;
 use Thinktomorrow\Chief\ManagedModels\Actions\SortModels;
 use Thinktomorrow\Chief\ManagedModels\Filters\Filters;
+use Thinktomorrow\Chief\ManagedModels\Events\ManagedModelsSorted;
 use Thinktomorrow\Chief\ManagedModels\Filters\Presets\HiddenFilter;
 use Thinktomorrow\Chief\Managers\Routes\ManagedRoute;
 
@@ -57,6 +58,8 @@ trait SortAssistant
             $this->managedModelClassInstance()->getKeyName(),
             $this->managedModelClassInstance()->getKeyType() == 'int',
         );
+
+        event(new ManagedModelsSorted($this->resource::resourceKey(), $request->indices));
 
         return response()->json([
             'message' => 'models sorted.',